requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
drf-yasg>=1.21
python-dotenv>=1.0
pytest>=8.0
pytest-django>=4.5
